Refactor server storages
