+++ conans/server/store/server_store.py
@@ -1,146 +1,34 @@
 from os.path import normpath, join, relpath
-
-from conans.errors import NotFoundException, ConanException
 from conans.model.ref import PackageReference, ConanFileReference
 from conans.paths import SimplePaths, EXPORT_FOLDER, PACKAGES_FOLDER
-from conans.server.revision_list import RevisionList
-
-REVISIONS_FILE = "revisions.txt"
 
 
 class ServerStore(SimplePaths):
 
-    def __init__(self, revisions_enabled, storage_adapter):
+    def __init__(self, storage_adapter):
         super(ServerStore, self).__init__(storage_adapter.base_storage_folder())
-        self._revisions_enabled = revisions_enabled
         self._storage_adapter = storage_adapter
 
-    # Methods to override some basics from paths to allow revisions paths (reading latest)
-    def conan(self, reference, resolve_latest=True):
-        reference = self.ref_with_rev(reference) if resolve_latest else reference
-        tmp = normpath(join(self.store, "/".join(reference)))
-        if reference.revision:
-            return join(tmp, reference.revision)
-        return tmp
+    def conan(self, reference):
+        return normpath(join(self.store, "/".join(reference)))
 
     def packages(self, reference):
-        reference = self.ref_with_rev(reference)
         return join(self.conan(reference), PACKAGES_FOLDER)
 
     def package(self, p_reference, short_paths=None):
-        p_reference = self._patch_package_ref(p_reference)
-        tmp = join(self.packages(p_reference.conan), p_reference.package_id)
-        if p_reference.revision:
-            return join(tmp, p_reference.revision)
-        return tmp
+        return join(self.packages(p_reference.conan), p_reference.package_id)
 
     def export(self, reference):
         return join(self.conan(reference), EXPORT_FOLDER)
 
-    # Methods to manage revisions
-    def get_last_revision(self, reference):
-        assert(isinstance(reference, ConanFileReference))
-        rev_file_path = self._recipe_revisions_file(reference)
-        return self._get_latest_revision(rev_file_path)
-
-    def get_last_package_revision(self, p_reference):
-        assert(isinstance(p_reference, PackageReference))
-        rev_file_path = self._package_revisions_file(p_reference)
-        return self._get_latest_revision(rev_file_path)
-
-    def update_last_revision(self, reference):
-        assert(isinstance(reference, ConanFileReference))
-        rev_file_path = self._recipe_revisions_file(reference)
-        self._update_last_revision(rev_file_path, reference)
-
-    def update_last_package_revision(self, p_reference):
-        assert(isinstance(p_reference, PackageReference))
-        rev_file_path = self._package_revisions_file(p_reference)
-        self._update_last_revision(rev_file_path, p_reference)
-
-    def _update_last_revision(self, rev_file_path, reference):
-        if self._storage_adapter.path_exists(rev_file_path):
-            rev_file = self._storage_adapter.read_file(rev_file_path,
-                                                       lock_file=rev_file_path + ".lock")
-            rev_list = RevisionList.loads(rev_file)
-        else:
-            rev_list = RevisionList()
-        if not reference.revision:
-            raise ConanException("Invalid revision for: %s" % reference.full_repr())
-        rev_list.add_revision(reference.revision)
-        self._storage_adapter.write_file(rev_file_path, rev_list.dumps(),
-                                         lock_file=rev_file_path + ".lock")
-
-    def _get_latest_revision(self, rev_file_path):
-        if self._storage_adapter.path_exists(rev_file_path):
-            rev_file = self._storage_adapter.read_file(rev_file_path,
-                                                       lock_file=rev_file_path + ".lock")
-            rev_list = RevisionList.loads(rev_file)
-            return rev_list.latest_revision()
-        else:
-            return None
-
-    def _recipe_revisions_file(self, reference):
-        recipe_folder = normpath(join(self._store_folder, "/".join(reference)))
-        return join(recipe_folder, REVISIONS_FILE)
-
-    def _package_revisions_file(self, p_reference):
-        tmp = normpath(join(self._store_folder, "/".join(p_reference.conan)))
-        revision = {None: ""}.get(p_reference.conan.revision, p_reference.conan.revision)
-        p_folder = join(tmp, revision, PACKAGES_FOLDER, p_reference.package_id)
-        return join(p_folder, REVISIONS_FILE)
-
     def get_conanfile_file_path(self, reference, filename):
-        reference = self.ref_with_rev(reference)
         abspath = join(self.export(reference), filename)
         return abspath
 
     def get_package_file_path(self, p_reference, filename):
-        p_reference = self._patch_package_ref(p_reference)
         p_path = self.package(p_reference)
         abspath = join(p_path, filename)
         return abspath
-
-    def ref_with_rev(self, reference):
-        if not self._revisions_enabled:
-            return reference.copy_without_revision()
-        if reference.revision:
-            return reference
-
-        latest = self.get_last_revision(reference)
-        if not latest:
-            raise NotFoundException("Recipe not found: '%s'" % reference.full_repr())
-
-        reference.revision = latest
-        return reference
-
-    def _patch_package_ref(self, p_reference):
-        if not self._revisions_enabled:
-            return p_reference.copy_without_revision()
-        if p_reference.revision:
-            return p_reference
-
-        reference = self.ref_with_rev(p_reference.conan)
-        ret = PackageReference(reference, p_reference.package_id)
-
-        latest = self.get_last_package_revision(ret)
-        if not latest:
-            raise NotFoundException("Package not found: '%s'" % str(p_reference))
-
-        ret.revision = latest
-        return ret
-
-    def update_recipe_revision(self, reference):
-        if not self._revisions_enabled:
-            return
-
-        self.update_last_revision(reference)
-
-    def update_package_revision(self, p_reference):
-        if not self._revisions_enabled:
-            return
-
-        self.update_last_package_revision(p_reference)
 
     def path_exists(self, path):
         return self._storage_adapter.path_exists(path)
@@ -154,7 +42,6 @@ class ServerStore(SimplePaths):
     def get_package_snapshot(self, p_reference):
         """Returns a {filepath: md5} """
         assert isinstance(p_reference, PackageReference)
-        p_reference = self._patch_package_ref(p_reference)
         path = self.package(p_reference)
         return self._get_snapshot_of_files(path)
 
@@ -166,20 +53,9 @@ class ServerStore(SimplePaths):
     # ######### DELETE (APIv1 and APIv2)
     def remove_conanfile(self, reference):
         assert isinstance(reference, ConanFileReference)
-        result = self._storage_adapter.delete_folder(self.conan(reference, resolve_latest=False))
-        if reference.revision:
-            rev_file_path = self._recipe_revisions_file(reference)
-            self._remove_revision(rev_file_path, reference.revision)
+        result = self._storage_adapter.delete_folder(self.conan(reference))
         self._storage_adapter.delete_empty_dirs([reference])
         return result
-
-    def _remove_revision(self, rev_file_path, revision):
-        rev_file = self._storage_adapter.read_file(rev_file_path,
-                                                   lock_file=rev_file_path + ".lock")
-        rev_list = RevisionList.loads(rev_file)
-        rev_list.remove_revision(revision)
-        self._storage_adapter.write_file(rev_file_path, rev_list.dumps(),
-                                         lock_file=rev_file_path + ".lock")
 
     def remove_packages(self, reference, package_ids_filter):
         assert isinstance(reference, ConanFileReference)
@@ -256,6 +132,12 @@ class ServerStore(SimplePaths):
         urls = self._storage_adapter.get_upload_urls(abs_paths, user)
         urls = self._relativize_keys(urls, relative_path)
         return urls
+
+    def update_recipe_reference(self, reference):
+        pass
+
+    def update_package_reference(self, p_reference):
+        pass
 
     @staticmethod
     def _relativize_keys(the_dict, basepath):
